Time: 0 ms (100%), Space: 9.5 MB (16.57%) - LeetHub

+++ 0410-split-array-largest-sum/0410-split-array-largest-sum.cpp
@@ -13,7 +13,7 @@ public:
 
     bool duable(long long mx, int m, vector<int>& nums) {
         long long sum = 0;
-        for(long long num : nums) {
+        for(int num : nums) {
             sum += num;
             if(sum > mx) {
                 sum = num;
